final final

+++ GoldenFeets/src/main/java/com/GoldenFeet/GoldenFeets/service/InventarioService.java
@@ -7,11 +7,20 @@ import java.util.List;
 import java.util.Optional;
 
 public interface InventarioService {
+
     List<Inventario> findAllInventarios();
+
     Optional<Inventario> findInventarioById(Long id);
+
     Inventario saveInventario(Inventario inventario);
+
     void deleteInventario(Long id);
-    Optional<Inventario> findByProductoId(Long productoId);
-    Inventario actualizarStock(Long productoId, int cantidad, String operacion); // "sumar" o "restar"
+
+    // --- CORRECCIÓN: Se cambió el parámetro de Long a Integer ---
+    Optional<Inventario> findByProductoId(Integer productoId);
+
+    // --- CORRECCIÓN: Se cambió el parámetro de Long a Integer ---
+    Inventario actualizarStock(Integer productoId, int cantidad, String operacion);
+
     Inventario crearInventarioInicial(Producto producto, int stockInicial);
 }
